requirements.txt
numpy
opencv-python-headless
Pillow
onnxruntime
rembg
requests